Uses one function to get GL instead of two
